Only apply system property overrides in settingsEvaluated/projectEvaluated
The system property override logic is moved to its own class.

This way, the original plugin behavior is preserved (and it is less confusing
for users who use theh plugin as a template for thier own plugins).
